requirements.txt
fastapi
uvicorn
pandas
strawberry-graphql
